Implement simple first-fit-decreasing algorithm
This should be good enough for what we need given the small numbers
involved.  There will be situations in which we could seat more people
by kicking out ones that already got assigned a table but we don't want
to do that.  Detecting the first group that cannot be seated, and will
thus end up on the waiting list, should be enough.

Future work: see if there are smaller groups further down the line that
will still fit in.